Refactor tab index and add Main component to forest
